+++ ck2/CHANGES.md
@@ -1,6 +1,6 @@
 # Prototyping phase
 
-## Dev
+## V0.7.7
    - added utils.get_current_date_time
    - added utils.assemble_cm_object1